% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annual runoff erosion power.R
\name{calculate_erosion_power}
\alias{calculate_erosion_power}
\title{计算年径流侵蚀功率}
\usage{
calculate_erosion_power(data0, Area)
}
\arguments{
\item{data0}{数据框，第一列为年份，第二列为月份，第三列及其往后为月流量值，m3/s。}

\item{Area}{数据框，第一列为流域编号，第二列为流域控制水文站名称，第三列为流域面积，km2。}
}
\value{
返回一个矩阵，第一列为年份，第二列及其往后为年径流侵蚀功率，m4/(s-1·km-2)。
}
\description{
这个函数计算给定流域的年径流侵蚀功率。
李占斌，鲁克新，李鹏，等.基于径流侵蚀功率的流域次暴雨产沙模型研究\link{C}.第六届全国泥沙基本理论研究学术讨论会论文集.郑州:黄河水利出版社，2005：54-59.
贾路，李占斌，于坤霞，等.基于径流侵蚀功率的长江典型流域能沙关系模型及改进\link{J}.农业工程学报,2024,40(05):128-140.
于坤霞，李天毅，贾路，等.黄土高原径流侵蚀功率输沙模型的改进\link{J}.农业工程学报,2024,40(10):107-116.
}
\examples{
data <- read.csv('./逐月平均流量.csv', sep =",", fileEncoding = "GB18030")
Area <- read.csv('./Area.csv', sep =",", T, fileEncoding = "GB18030")
E <- calculate_erosion_power(data, Area)
}
